dtypes corrections

+++ dbt_project/models/staging/stg_devices.sql
@@ -10,9 +10,9 @@ with source as (
 devices as (
 
     select
-        trim(id) as device_id,
-        trim(type) as device_type,
-        trim(store_id) as store_id
+        trim(id)::numeric as device_id,
+        trim(type)::numeric as device_type,
+        trim(store_id)::numeric as store_id
 
     from source
