+++ dbt_project/models/staging/stg_stores.sql
@@ -10,14 +10,14 @@ with source as (
 stores as (
 
     select
-        trim(id) as store_id,
+        trim(id)::numeric as store_id,
         trim(name) as store_name,
         trim(address) as address,
         trim(city) as city,
         trim(country) as country,
         to_timestamp(created_at, 'MM/DD/YYYY HH24:MI:SS') as created_at,
         trim(typology) as typology,
-        trim(customer_id) as customer_id
+        trim(customer_id)::numeric as customer_id
 
     from source
